backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-docx
python-multipart
python-slugify
aiosqlite
